// src/gen/bikehub/model/bicycle.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BicycleCategory } from './bicycleCategory';
import { BicycleStatus } from './bicycleStatus';


export interface Bicycle {
    id: string;
    currentParkingLotName: string;
    categories?: Set<BicycleCategory>;
    equipments?: Array<string>;
    chassisId: string;
    status: BicycleStatus;
    brand: string;
    model: string;
    totalRentTimeInSeconds?: number;
    hourlyPrice: number;
    registeredAt: string;
}