// src/gen/bikehub/model/paymentType.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum PaymentType {

    Cash = 'CASH',

    CreditCard = 'CREDIT_CARD',

    DebitCard = 'DEBIT_CARD'
}