// src/gen/bikehub/model/rental.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RentalStatus } from './rentalStatus';


export interface Rental {
    id: string;
    bicycleId: string;
    pickUpParkingLotName: string;
    returnParkingLotName?: string;
    status: RentalStatus;
    createdAt: string;
    startedAt?: string;
    finishedAt?: string;
    totalPrice?: number;
    mileage?: number;
}