// src/gen/bikehub/model/rentalStatus.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum RentalStatus {

    Created = 'CREATED',

    InProgress = 'IN_PROGRESS',

    Finished = 'FINISHED',

    Payed = 'PAYED'
}